+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 2021-10-02
+- Establish basic status text in the client.
 # 2021-10-01
 - Restarting this project with a focus on the ability to play correspondence games against other players.
 - Cleared up instructions for running the codebase.